HBX-1952: Move class 'DatabaseReader' to new package 'org.hibernate.tool.internal.reveng.reader'
